Renaming Common.cs to Timekeeper.cs to separate Timekeeper's "Common" 	methods from Toolbox's Common methods. 	----------------------------------------------------------------------- 	TESTING 	- n/a 	----------------------------------------------------------------------- 	DETAILS 	- Straightforward 	----------------------------------------------------------------------- 	Project...: Timekeeper 	Components: Common 	Target....: 2.3 	Ticket....: n/a
[git-p4: depot-paths = "//depot/timekeeper/": change = 209]

+++ timekeeper/Timekeeper.cs
@@ -28,7 +28,7 @@ namespace Timekeeper
             return DateTime.Now.ToString(Common.DATETIME_FORMAT);
         }
 
-        public static string FormatSeconds(int seconds)
+        public static string FormatSeconds(long seconds)
         {
             TimeSpan t = TimeSpan.FromSeconds(seconds);
             return string.Format("{0:D2}:{1:D2}:{2:D2}",
